feat(offline): create instance of ApolloClient is quite complex so extracted factory method
