+++ packages/falcon-client/src/service/index.js
@@ -1,2 +1,3 @@
 export * from './ApolloClient';
+export * from './ApolloClientWeb';
 export * from './apolloStateToObject';
